Edited the smtp client - was incorrectly used in the using. Then added the ssl.
Had to move the send() to EmailCredentials because smtp client cannot be read outside of the using inside SetUpAndSendEmail() because it gets disposed.

Email will now send but it uses a lot of memory as its done 1 by 1 and the images are really big.

Need to improve by adding the emails to the To field (maybe) and decrease the size of the image sent.